Revert "planner: move code of rebuilding cached plans into a separate file (#53163)"
This reverts commit 87247c113cde36f3dfdec7cd7836cb5129431094.
