+++ pkg/planner/core/plan_cache_utils.go
@@ -491,16 +491,6 @@ func (*PlanCacheQueryFeatures) Leave(in ast.Node) (out ast.Node, ok bool) {
 	return in, true
 }
 
-// PointGetExecutorCache caches the PointGetExecutor to further improve its performance.
-// Don't forget to reset this executor when the prior plan is invalid.
-type PointGetExecutorCache struct {
-	ColumnInfos any
-	// Executor is only used for point get scene.
-	// Notice that we should only cache the PointGetExecutor that have a snapshot with MaxTS in it.
-	// If the current plan is not PointGet or does not use MaxTS optimization, this value should be nil here.
-	Executor any
-}
-
 // PlanCacheStmt store prepared ast from PrepareExec and other related fields
 type PlanCacheStmt struct {
 	PreparedAst *ast.Prepared
@@ -508,7 +498,16 @@ type PlanCacheStmt struct {
 	VisitInfos  []visitInfo
 	Params      []ast.ParamMarkerExpr
 
-	PointGet PointGetExecutorCache
+	// To further improve the performance of PointGet, cache execution info for PointGet directly.
+	// Use any to avoid cycle import.
+	// TODO: caching execution info directly is risky and tricky to the optimizer, refactor it later.
+	PointGet struct {
+		ColumnInfos any
+		// Executor is only used for point get scene.
+		// Notice that we should only cache the PointGetExecutor that have a snapshot with MaxTS in it.
+		// If the current plan is not PointGet or does not use MaxTS optimization, this value should be nil here.
+		Executor any
+	}
 
 	// below fields are for PointGet short path
 	SchemaVersion int64
@@ -722,24 +721,4 @@ func isSafePointGetPath4PlanCacheScenario3(path *util.AccessPath) bool {
 		}
 	}
 	return true
-}
-
-// parseParamTypes get parameters' types in PREPARE statement
-func parseParamTypes(sctx sessionctx.Context, params []expression.Expression) (paramTypes []*types.FieldType) {
-	paramTypes = make([]*types.FieldType, 0, len(params))
-	for _, param := range params {
-		if c, ok := param.(*expression.Constant); ok { // from binary protocol
-			paramTypes = append(paramTypes, c.GetType())
-			continue
-		}
-
-		// from text protocol, there must be a GetVar function
-		name := param.(*expression.ScalarFunction).GetArgs()[0].String()
-		tp, ok := sctx.GetSessionVars().GetUserVarType(name)
-		if !ok {
-			tp = types.NewFieldType(mysql.TypeNull)
-		}
-		paramTypes = append(paramTypes, tp)
-	}
-	return
 }
